+++ include/wrench/services/compute/multihost_multicore/MultihostMulticoreComputeService.h
@@ -76,7 +76,7 @@ namespace wrench {
                                          const bool supports_pilot_jobs,
                                          const std::set<std::tuple<std::string, unsigned long, double>> compute_resources,
                                          std::map<std::string, std::string> plist = {},
-                                         int scratch_size = 0);
+                                         double scratch_size = 0);
 
         // Public Constructor
         MultihostMulticoreComputeService(const std::string &hostname,
@@ -84,7 +84,7 @@ namespace wrench {
                                          bool supports_pilot_jobs,
                                          const std::set<std::string> compute_hosts,
                                          std::map<std::string, std::string> plist = {},
-                                         int scratch_size = 0);
+                                         double scratch_size = 0);
 
 
         /***********************/
